Make entrypoint script smarter

+++ docker-entrypoint.sh
@@ -3,6 +3,11 @@
 SFTP_PIDFILE="/tmp/sftp-server.pid"
 
 if [ "${*}" = "yarn start" ]; then
+    if [ ! -f "/srv/daemon/config/core.json" ]; then
+        echo ">>> ERROR: configuration is not present"
+        exit 1
+    fi
+
     # Start SFTP server
     /srv/daemon/sftp-server &
     echo "${!}" > "${SFTP_PIDFILE}"
@@ -10,7 +15,7 @@ if [ "${*}" = "yarn start" ]; then
     yarn start
 
     # Stop SFTP server
-    if [ -f "${SFTP_PIDFILE}" ]; then
+    if [ -f "${SFTP_PIDFILE}" ] && (kill -0 "$(cat "${SFTP_PIDFILE}")"); then
         kill -15 "$(cat "${SFTP_PIDFILE}")"
     fi
 else
